Reparación y testeo de todos los querys con casos de prueba
se corrigió errores con excepciones cuando el valor retorna null, entre otros

+++ src/main/java/pe/edu/upc/visually_impaired/controllers/Rutas_segurasController.java
@@ -66,7 +66,7 @@ public class Rutas_segurasController {
             dto.setPunto_origen(columna[0]);
             dto.setPunto_destino(columna[1]);
             dto.setDescripcion(columna[2]);
-            dto.setTiempo_promedio_navegacion_minutos(Integer.parseInt(columna[3]));
+            dto.setTiempo_promedio_navegacion_minutos(Float.parseFloat(columna[3]));
             dtoLista.add(dto);
         }
         return dtoLista;
